Backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
python-jose[cryptography]==3.3.0
bcrypt==4.0.1
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
httpx==0.25.2
pandas>=2.0.0
scikit-learn>=1.3.0
numpy>=1.24.0
joblib>=1.3.0
python-dateutil>=2.8.0
email-validator>=2.1.0
